Add automatic recovery of *.tmp files during synchronization.
Added an automatic recovery routine to ImportUtil.cs:Import().

On file open failure, automatically searches for a *.tmp file and will
attempt to rename it as the target file *IF* the target file actually
does not exist.

This prevents an interrupted WebDAV DELETE/MOVE pair from completely
breaking synchronization (for example, using triggers).