More targeted fix for KeePass freezing single-thread WMs with xsel.
This fix adds a new parameter for RunConsoleApp which allows the caller to
specify that it should not wait for the subprocess to exit synchronously. The
Unix-specific xsel invocation is then modified to use this new method, which
prevents xsel from blocking the main UI thread and under some circumstances
locking up the window-manager when run under Mono.

This fix should be generally applicable to any uses of xsel, but does not change
the behavior of any other RunConsoleApp uses.