More targeted fix for KeePass freezing single-thread WMs.
This fix adds a new parameter for RunConsoleApp which allows the caller
to specify that it should not wait for the subprocess to exit
synchronously. The Unix-specific xsel invocation is then modified to use
this new method, which prevents xsel from blocking the main UI thread.

This is generally applicable to all uses of xsel.